Nimber extends ImpartialGame
git-svn-id: svn://svn.code.sf.net/p/cgsuite/code/cgsuite/branches/devel@510 0138b134-6283-4d47-a013-9bc70bc83b06

+++ lib/core/src/main/scala/org/cgsuite/lang/Ops.scala
@@ -153,7 +153,7 @@ object Ops {
   }
 
   val MakeNimber = UnOp("nim") {
-    case x: SmallInteger => Nimber(x.intValue)
+    case x: SmallInteger => Nimber(x)
   }
 
   val MakeUpMultiple = BinOp("up") {
